+++ CHANGES.rst
@@ -7,6 +7,7 @@ Version 0.2
 
 - Add metadata from netCDF file to returned image.
 - Add option to return data as 1D arrays.
+- Add image to time series conversion and time series reading interface.
 
 Version 0.1
 ===========